+++ internal/handlers/order.go
@@ -72,7 +72,7 @@ func (h *Handler) CreateOrder(w http.ResponseWriter, r *http.Request) {
 type GetOrdersResponseItem struct {
 	Number     string    `json:"number"`
 	Status     string    `json:"status"`
-	Accrual    *int64    `json:"accrual,omitempty"`
+	Accrual    *float64  `json:"accrual,omitempty"`
 	UploadedAt time.Time `json:"uploaded_at"`
 }
